refactor: adjusts file names and folder structures for improved readability and maintainability
