+++ index.html
@@ -1,13 +1,14 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/logo.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./src/index.css" />
     <title>FlowHub - Produtividade</title>
   </head>
   <body>
-    <div id="root"></div>
+    <div id="root" />
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
