+++ src/components/syntax-highlighter.tsx
