+++ src/components/weather-widget.tsx
