+++ src/pages/code-snippets.tsx
@@ -8,7 +8,7 @@ import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@
 import { useToast } from "@/hooks/use-toast";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { FileText, Clipboard, Plus, Trash2, Code, Search, Filter } from "lucide-react";
-import CodeHighlighter from "@/components/SyntaxHighlighter";
+import CodeHighlighter from "@/components/syntax-highlighter";
 import { motion } from "framer-motion";
 import { Badge } from "@/components/ui/badge";
