+++ src/pages/tech-news.tsx
